implemented at simplification rules

+++ condition_extractor/src/GenericDominatorTy.cpp
@@ -3,32 +3,35 @@
 
 #include "SVF-FE/LLVMUtil.h"
 
+// FLAVIO: I intentionally left the burned of creating the Dom relation in a
+// separate function. I did not want to do everything in the constructor 
+void GenericDominatorTy::createDom() {
 
-void GenericDominatorTy::createDom(GenericDominatorTy *dom) {
-
-    if (dom == nullptr) {
-        outs() << "[ERROR] Dom pointer is NULL!\n";
-        exit(1);
+    if (is_created) {
+        outs() << "[ERROR] alrady created!\n";
+        return;
     }
 
     outs() << "[INFO] Running pruneUnreachableFunctions()\n";
-    pruneUnreachableFunctions(dom);
+    this->pruneUnreachableFunctions();
     outs() << "[INFO] Running buildPhiFun()\n";
-    buildPhiFun(dom);
+    this->buildPhiFun();
     outs() << "[INFO] Running inferSubGraph()\n";
-    inferSubGraph(dom);
+    this->inferSubGraph();
     outs() << "[INFO] Running buildR()\n";
-    buildR(dom);
+    this->buildR();
     outs() << "[INFO] Running buildDom()\n";
-    dom->buildDom();
+    this->buildDom();
     outs() << "[INFO] Running restoreUnreachableFunctions()\n";
-    restoreUnreachableFunctions(dom);
+    this->restoreUnreachableFunctions();
+
+    is_created = true;
 
 }
 
-void GenericDominatorTy::restoreUnreachableFunctions(GenericDominatorTy* dom) {
+void GenericDominatorTy::restoreUnreachableFunctions() {
     // outs() << "[INFO] Restore eddges\n";
-    for (auto edge: dom->getDumpedEdge()) {
+    for (auto edge: this->getDumpedEdge()) {
         // outs() << edge->toString() << "\n";
         edge->getDstNode()->addIncomingEdge(edge);
         edge->getSrcNode()->addOutgoingEdge(edge);
@@ -39,30 +42,30 @@ void GenericDominatorTy::restoreUnreachableFunctions(GenericDominatorTy* dom) {
 }
 
 
-void GenericDominatorTy::buildPhiFun(GenericDominatorTy *dom) {
+void GenericDominatorTy::buildPhiFun() {
 
-    SVFModule *svfModule = dom->getModule();
-    ICFG *icfg = dom->getICFG();
+    SVFModule *svfModule = this->getModule();
+    ICFG *icfg = this->getICFG();
 
     PHIFun phi;
     PHIFunInv phi_inv;
 
     getPhiFunction(svfModule, icfg, &phi, &phi_inv);    
 
-    dom->setPhi(phi);
-    dom->setPhiInv(phi_inv);
+    this->setPhi(phi);
+    this->setPhiInv(phi_inv);
 
     // NOTE: to uncomment for debug
-    // dom->printPhiFunction();
-    // dom->printPhiInvFunction();
+    // this->printPhiFunction();
+    // this->printPhiInvFunction();
     // exit(1);
 
 }
 
-void GenericDominatorTy::buildR(GenericDominatorTy* dom) {
+void GenericDominatorTy::buildR() {
 
-    SVFModule *svfModule = dom->getModule();
-    ICFG* icfg = dom->getICFG();
+    SVFModule *svfModule = this->getModule();
+    ICFG* icfg = this->getICFG();
 
     SVF::SVFModule::llvm_iterator it = svfModule->llvmFunBegin();SVF::SVFModule::llvm_iterator eit = svfModule->llvmFunEnd();
 
@@ -86,7 +89,7 @@ void GenericDominatorTy::buildR(GenericDominatorTy* dom) {
                 // outs() << "it_fun_exit: " << (*it_fun_exit)->toString() << "\n";
                 ret_edge = (RetCFGEdge*)(*it_fun_exit);
                 
-                dom->addR(ret_edge);
+                this->addR(ret_edge);
             }
         }
 
@@ -97,26 +100,26 @@ void GenericDominatorTy::buildR(GenericDominatorTy* dom) {
                 // outs() << "it_fun_exit: " << (*it_fun_exit)->toString() << "\n";
                 call_edge = (CallCFGEdge*)(*it_fun_entry);
                 
-                dom->addC(call_edge);
+                this->addC(call_edge);
             }
         }
 
     }
 
     // NOTE: for debug
-    // dom->printR();
-    // dom->printC();
+    // this->printR();
+    // this->printC();
     // exit(1);
 }
 
-void GenericDominatorTy::pruneUnreachableFunctions(GenericDominatorTy *dom) {
+void GenericDominatorTy::pruneUnreachableFunctions() {
 
-    assert(dom->getEntryNode() && "We need an entry block!");
+    assert(this->getEntryNode() && "We need an entry block!");
 
-    const SVFFunction *main_fun = dom->getEntryNode()->getFun();
+    const SVFFunction *main_fun = this->getEntryNode()->getFun();
 
-    SVFModule *svfModule = dom->getModule();
-    PTACallGraph* callgraph = dom->getPTACallGraph();
+    SVFModule *svfModule = this->getModule();
+    PTACallGraph* callgraph = this->getPTACallGraph();
 
     SVF::SVFModule::llvm_iterator it, eit;
 
@@ -145,7 +148,7 @@ void GenericDominatorTy::pruneUnreachableFunctions(GenericDominatorTy *dom) {
             if (functions_done.find(fun) != functions_done.end())
                 continue;
 
-            FunEntryICFGNode *fun_entry = dom->getICFG()->getFunEntryICFGNode(fun);
+            FunEntryICFGNode *fun_entry = this->getICFG()->getFunEntryICFGNode(fun);
             FunEntryICFGNode *entry_called;
             FunExitICFGNode *exit_called;
             PTACallGraphNode *node_callee, *node_called;
@@ -173,7 +176,7 @@ void GenericDominatorTy::pruneUnreachableFunctions(GenericDominatorTy *dom) {
 
                         // have to select the callee node in ICFG, before it was
                         // the calle in CF
-                        entry_called = dom->getICFG()
+                        entry_called = this->getICFG()
                                         ->getFunEntryICFGNode(fun_called);
                         ICFGNode::const_iterator it3 = 
                                         entry_called->InEdgeBegin();
@@ -194,7 +197,7 @@ void GenericDominatorTy::pruneUnreachableFunctions(GenericDominatorTy *dom) {
                             tmp_dumped_edges.insert(edge_to_remove);
                         }
 
-                        exit_called = dom->getICFG()->getFunExitICFGNode(fun_called);
+                        exit_called = this->getICFG()->getFunExitICFGNode(fun_called);
                         if (exit_called->hasOutgoingEdge()) {
                             edge_to_remove = nullptr;
                             it3 = exit_called->OutEdgeBegin();
@@ -222,7 +225,7 @@ void GenericDominatorTy::pruneUnreachableFunctions(GenericDominatorTy *dom) {
                     edge->getDstNode()->removeIncomingEdge(edge);
                     edge->getSrcNode()->removeOutgoingEdge(edge);
                     // dumped_edges.insert(edge);
-                    dom->addDumpedEdge(edge);
+                    this->addDumpedEdge(edge);
                 }
 
                 functions_done.insert(fun);
@@ -230,7 +233,7 @@ void GenericDominatorTy::pruneUnreachableFunctions(GenericDominatorTy *dom) {
         }
     }
 
-    // dom->getICFG()->dump("icfg_pruned");
+    // this->getICFG()->dump("icfg_pruned");
 
     // outs() << "FUNCTIONS:\n";
     
@@ -242,9 +245,9 @@ void GenericDominatorTy::pruneUnreachableFunctions(GenericDominatorTy *dom) {
 }
 
 
-void GenericDominatorTy::inferSubGraph(GenericDominatorTy *dom) {
+void GenericDominatorTy::inferSubGraph() {
 
-    FunEntryICFGNode *entry_node = dom->getEntryNode();
+    FunEntryICFGNode *entry_node = this->getEntryNode();
 
     assert(entry_node && "We need an entry block!");
 
@@ -274,7 +277,7 @@ void GenericDominatorTy::inferSubGraph(GenericDominatorTy *dom) {
                     working.push(dst);
                 }
                 else if(auto call_edge = SVFUtil::dyn_cast<CallCFGEdge>(edge)) {
-                    ICFGEdge *next_ret = dom->getPhi(call_edge);
+                    ICFGEdge *next_ret = this->getPhi(call_edge);
                     ICFGNode *dst_r = next_ret->getDstNode();
                     
                     working.push(dst);
@@ -290,10 +293,10 @@ void GenericDominatorTy::inferSubGraph(GenericDominatorTy *dom) {
 
     }
 
-    dom->setRelevantNodes(visited);
+    this->setRelevantNodes(visited);
 
-    // outs() << "[INFO] Interesting nodes " << dom->getTotRelevantNodes() << "\n";
-    // ICFG* icfg = dom->getICFG();
+    // outs() << "[INFO] Interesting nodes " << this->getTotRelevantNodes() << "\n";
+    // ICFG* icfg = this->getICFG();
     // outs() << "[INFO] All nodes " << icfg->getTotalNodeNum() << "\n";
     // for (auto n: visited)
     //     outs() << n->toString() << "\n";
@@ -315,6 +318,10 @@ GenericDominatorTy::GenericDominatorTy(BVDataPTAImpl* a_point_to)
 }
 
 bool GenericDominatorTy::dominates(ICFGNode *a, ICFGNode *b) {
+    if (!is_created) {
+        outs() << "[ERROR] " << getDomName() << " not created yet!\n";
+        exit(1);
+    }
     ICFGNodeSet dominators_b = getDom(b);
     return dominators_b.find(a) != dominators_b.end();
 }
@@ -324,6 +331,11 @@ bool GenericDominatorTy::dominates(ICFGNode *a, ICFGNode *b) {
  */
 void GenericDominatorTy::dumpTransRed(const std::string& file, bool simple)
 {
+    if (!is_created) {
+        outs() << "[ERROR] " << getDomName() << " not created yet!\n";
+        exit(1);
+    }
+
     outs() << "[INFO] Dom covering " << getTotRelevantNodes() << "\n";
     outs() << "[INFO] Running transient reduction...\n";
     outs() << "[INFO] This might thake a while..." 
@@ -333,7 +345,13 @@ void GenericDominatorTy::dumpTransRed(const std::string& file, bool simple)
     GraphPrinter::WriteGraphToFile(SVFUtil::outs(), file, this, simple);
 }
 
-void GenericDominatorTy::dumpDom(const std::string& file) {
+void GenericDominatorTy::dumpDom(const std::string& file)
+{
+    if (!is_created) {
+        outs() << "[ERROR] " << getDomName() << " not created yet!\n";
+        exit(1);
+    }
+
     outs() << "[INFO] Dom covering " << getTotRelevantNodes() << "\n";
     outs() << "[INFO] This might thake a while..." 
             << "if too long, kill the process\n";
